video-duration-api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
environs==10.3.0
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi_cors==0.0.6
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.60.1
h11==0.14.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
marshmallow==3.20.2
moviepy==1.0.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
